Add configuration interface and implement their methods. Make improvements for logger. Split command from main.go
