+++ cmd/bpfink/consts.go
@@ -1,0 +1,10 @@
+package bpfink
+
+const (
+	// DefaultConfigFile default New file location
+	DefaultConfigFile = "/etc/bpfink.toml"
+	// DefaultDatabase default database file location
+	DefaultDatabase       = "/var/lib/bpfink.db"
+	puppetFileColumnCount = 2
+	keySize               = 16
+)
